First draft with the new orchestrator version
Scripts generated using the tool, passing in CI but needs some improvements in order to solve the problems

+++ retorchfiles/scripts/tjoblifecycles/tjob-teardown.sh
@@ -1,32 +1,37 @@
 #!/bin/bash
+# The tjob-teardown.sh script provides all the necessary commands to tear-down each TJob's Resources after the
+# test execution has ended. It stores the container logs of the current TJob, tear down the containers
+# and execute the custom commands provided in the custom-tjob-teardown file.
+
 set -e
+
+if [ "$#" -ne 2 ]; then
+    "$SCRIPTS_FOLDER/printLog.sh" "ERROR" "TJob-tear-down" "Usage: $0 <TJobName> <Stage>"
+    exit 1
+fi
 
 # Execute the script to write timestamp
 "$SCRIPTS_FOLDER/writetime.sh" "$2" "$1"
-
+"$SCRIPTS_FOLDER/printLog.sh" "DEBUG" "$1-tear-down" "Starting the TJob tear-down"
 # Store docker logs
-DIRECTORY_PATH="$WORKSPACE/target/containerlogs/$1"
-
-if [ ! -d "$DIRECTORY_PATH" ]; then
-  echo "Directory for storing logs doesnt exist creating..."
-  mkdir -p $DIRECTORY_PATH
-fi
-
-for CONTAINER_NAME in $(docker ps -a --format "{{.Names}}" --filter Name=$1); do
-  echo "Storing lof con container $CONTAINER_NAME in TJob $1"
-  docker logs $CONTAINER_NAME &>"$DIRECTORY_PATH/$CONTAINER_NAME.log"
-done
+"$SCRIPTS_FOLDER/storeContainerLogs.sh" "$1"
 
 # Change to SUT location
 cd "$SUT_LOCATION"
 
 # Tear down Docker containers and volumes
-echo "Tearing down Docker containers and volumes for TJOB $1"
+"$SCRIPTS_FOLDER/printLog.sh" "DEBUG" "$1-tear-down" "Tearing down Docker containers and volumes for TJOB $1"
 docker compose -f docker-compose.yml --env-file "$WORKSPACE/retorchfiles/envfiles/$1.env" --ansi never -p "$1" down --volumes
 
 # Return to the original working directory
 cd "$WORKSPACE"
 
+# START Custom Set-up commands
+"$SCRIPTS_FOLDER/printLog.sh" "DEBUG" "$1-set-up" "Start executing custom commands"
+echo "This TJOB-teardown dont have any kind of specific commands"
+"$SCRIPTS_FOLDER/printLog.sh" "DEBUG" "$1-set-up" "End executing custom commands"
+# END Custom Set-up commands
 
+"$SCRIPTS_FOLDER/printLog.sh" "DEBUG" "$1-tear-down" "Tear-down ended"
 # Execute the script to write timestamp again
 "$SCRIPTS_FOLDER/writetime.sh" "$2" "$1"
